+++ Sources/KeyboardKit/_Deprecated/Image+Deprecated.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+
+public extension Image {
+
+    @available(*, deprecated, renamed: "keyboardShiftCapslockActive")
+    static var keyboardShiftCapslocked: Image {
+        .keyboardShiftCapslockActive
+    }
+}